Arreglar errors de pagament i pujada de versió (#54)
* Add user-based filtering for payments in history feature

- Update `PaymentRepository` interface and `FirebasePaymentRepository` implementation to require a `UserId` when fetching payments by month.
- Update the Firestore query in `FirebasePaymentRepository` to filter by the `user_id` field.
- Refactor `HistoryViewModel` to combine `currentDate` and `authRepository.getCurrentUserFlow()`, ensuring payments are fetched for the currently authenticated user.

* Restrict app orientation to portrait on Android and iOS

- Update `MainActivity` in `AndroidManifest.xml` to set `android:screenOrientation="portrait"`.
- Modify the iOS project configuration to restrict `UISupportedInterfaceOrientations` to `UIInterfaceOrientationPortrait`, removing device-specific overrides for iPhone and iPad.
- Clean up minor indentation in the iOS `Info.plist`.

* Implement payment calculation logic and update app versions

- Refactor `Volunteer.calculateTotalToPay()` to compute totals based on meal counts (excluding breakfast) and overnight stays.
- Add `Meal.Breakfast` to the `Meal` enum and define a `MEAL_PRICE` constant in the `Volunteer` model.
- Update `ReservationFormViewModel` to handle the updated payment calculation return type.
- Increment Android `versionName` and iOS `MARKETING_VERSION` to `1.0.11`.

---------

+++ features/calendar/src/commonMain/kotlin/com/casavirupa/voluntariat/features/calendar/viewmodels/ReservationFormViewModel.kt
@@ -318,7 +318,7 @@ class ReservationFormViewModel(
                                 paymentRepository.addPayment(
                                     userId = user.id,
                                     yearMonth = date.value!!.toYearMonth(),
-                                    amount = volunteer.calculateTotalToPay(),
+                                    amount = volunteer.calculateTotalToPay().toDouble(),
                                 )
                             }.onSuccess {
                                 navigateBack()
